requirements.txt
# UI
streamlit>=1.36.0

# ENV
python-dotenv>=1.0.1

# LLM / Azure OpenAI
openai>=1.40.0

# Data / models
pydantic>=2.7.0

# Vector RAG
chromadb>=0.5.5
sentence-transformers>=3.0.1

# Graph
networkx>=3.3
neo4j>=5.22.0

# PDF
pypdf>=4.2.0
pyyaml>=6.0.1
jsonschema>=4.22.0
